// Client/VariantConfigurationWeb/src/app/models/order-accesories.model.ts
export interface OrderAccessoriesModel {
    TabletHolderId?: number;
    TabletHolderQuantity: number;
    PhoneHolderId?: number;
    PhoneHolderQuantity: number;
    MediaShelfId?: number;
    MediaShelfQuantity: number;
    SeatPostId?: number;
    SeatPostQuantity: number;
    HandlebarPostId?: number;
    HandlebarPostQuantity: number;
    StagesDumbbellHolderId?: number;
    StagesDumbbellHolderQuantity: number;
    AerobarId?: number;
    AerobarQuantity: number;
    PlatesOneToSixtyId?: number;
    PlatesOneToSixtyQuntity: number;
    PlatesSixtyOneToHundredId?: number;
    PlatesSixtyOneToHundredQuantity: number;
    PlatesOneToEightyId?: number;
    PlatesOneToEightyQuantity: number;
    PlatesFiftyPeacesId?: number;
    PlatesFiftyPeacesQuantity: number;
    PlatesOneToThirtyId?: number;
    PlatesOneToThirtyQuantity: number;
}