// Client/VariantConfigurationWeb/src/app/models/accessory-quantity.model.ts
export interface AccessoryQuantityModel {
    tabletHolderQuantity: number;
    mediaShelfQuantity: number;
    handlebarPostQuantity: number;
    seatPostQuantity: number;
    phoneHolderQuantity: number;
    stagesDumbbellHolderQuantity: number;
    aerobarQuantity: number;
    sbnpOneToSixtyQuantity: number;
    sbnpSixtyOneToHundredQuantity: number;
    sbnpOneToEightyQuantity: number;
    sbnpOneStagesLogoQuantity: number;
    sbnpOneToThirtyQuantity: number;
}