// Client/VariantConfigurationWeb/src/app/models/accessory-price.model.ts
export interface AccessoryPriceModel {
    tabletHolderTotalPrice: number;
    mediaShelfTotalPrice: number;
    handlebarPostTotalPrice: number;
    seatPostTotalPrice: number;
    phoneHolderTotalPrice: number;
    stagesDumbbellHolderTotalPrice: number;
    aerobarTotalPrice: number;
    sbnpOneToSixtyTotalPrice: number;
    sbnpSixtyOneToHundredTotalPrice: number;
    sbnpOneToEightyTotalPrice: number;
    sbnpOneStagesLogoTotalPrice: number;
    sbnpOneToThirtyTotalPrice: number;

}